MapContextEditor: Extract MapContextLayerRepoForm component.
